+++ autocomplete-angular/src/app/autocomplete/index.ts
@@ -3,6 +3,6 @@ export { AutocompleteController } from './autocomplete-controller';
 export { AutocompleteTypeset, AutocompleteTypeProvider } from './autocomplete-type-provider';
 export { AutocompleteCreateComponent } from './autocomplete-create-component';
 export { AutocompleteListComponent } from './autocomplete-list-component';
-export { AutocompleteTextFunction, AutocompleteResolveData, AutocompleteResolveFunction, Autocomplete, AutocompleteContainer } from './autocomplete';
+export { AutoAssignMode, AutocompleteTextFunction, AutocompleteResolveData, AutocompleteResolveFunction, Autocomplete, AutocompleteContainer } from './autocomplete';
 export { AutocompleteQueryProcessor, AutocompleteQueryMediator, BindQueryProcessorFunction, InputChangedFunction } from './autocomplete-query-mediator';
 export { AutocompleteModule } from './autocomplete.module';